+++ interface/forms/fee_sheet/new.php
@@ -47,6 +47,7 @@ require_once("$srcdir/options.inc.php");
 require_once("$srcdir/formdata.inc.php");
 require_once("$srcdir/log.inc");
 require_once("$srcdir/headers.inc.php");
+require_once("functions/fee_sheet_inc.php");
 
 // For logging checksums set this to true.
 define('CHECKSUM_LOGGING', true);
@@ -57,423 +58,6 @@ $usbillstyle = $GLOBALS['ippf_specific'] ? " style='display:none'" : "";
 // This may be an error message or warning that pops up when the form is loaded.
 $alertmsg = '';
 
-function alphaCodeType($id) {
-  global $code_types;
-  foreach ($code_types as $key => $value) {
-    if ($value['id'] == $id) return $key;
-  }
-  return '';
-}
-
-// Helper function for creating drop-lists.
-function endFSCategory() {
-  global $i, $last_category, $FEE_SHEET_COLUMNS;
-  if (! $last_category) return;
-  echo "   </select>\n";
-  echo "  </td>\n";
-  if ($i >= $FEE_SHEET_COLUMNS) {
-    echo " </tr>\n";
-    $i = 0;
-  }
-}
-
-// Generate JavaScript to build the array of diagnoses.
-function genDiagJS($code_type, $code) {
-  global $code_types;
-  if ($code_types[$code_type]['diag']) {
-    echo "diags.push('" . attr($code_type) . "|" . attr($code) . "');\n";
-  }
-}
-
-// For IPPF only.  Returns 0 = none, 1 = nonsurgical, 2 = surgical.
-//
-function contraceptionClass($code_type, $code) {
-  global $code_types;
-  if (!$GLOBALS['ippf_specific']) return 0;
-  $contra = 0;
-  // Get the related service codes.
-  $codesrow = sqlQuery("SELECT related_code FROM codes WHERE " .
-    "code_type = ? " .
-    " AND code = ? LIMIT 1", array($code_types[$code_type]['id'],$code) );
-  if (!empty($codesrow['related_code']) && $code_type == 'MA') {
-    $relcodes = explode(';', $codesrow['related_code']);
-    foreach ($relcodes as $relstring) {
-      if ($relstring === '') continue;
-      list($reltype, $relcode) = explode(':', $relstring);
-      if ($reltype !== 'IPPF') continue;
-      if      (preg_match('/^11....110/'    , $relcode)) $contra |= 1;
-      else if (preg_match('/^11....999/'    , $relcode)) $contra |= 1;
-      else if (preg_match('/^112152010/'    , $relcode)) $contra |= 1;
-      else if (preg_match('/^11317[1-2]111/', $relcode)) $contra |= 1;
-      else if (preg_match('/^12118[1-2].13/', $relcode)) $contra |= 2;
-      else if (preg_match('/^12118[1-2]999/', $relcode)) $contra |= 2;
-    }
-  }
-  return $contra;
-}
-# gets the provider from the encounter file , or from the logged on user or from the patient file
-function findProvider() {
-  global $encounter, $pid;
-  $find_provider = sqlQuery("SELECT provider_id FROM form_encounter " .
-        "WHERE pid = ? AND encounter = ? " .
-        "ORDER BY id DESC LIMIT 1", array($pid,$encounter) );
-  $providerid = $find_provider['provider_id'];
-  if($providerid == 0) {
-   $get_authorized = $_SESSION['userauthorized'];
-   if($get_authorized ==1) {
-      $providerid = $_SESSION[authUserID];
-   }
-  }
-  if($providerid == 0) {
-    $find_provider = sqlQuery("SELECT providerID FROM patient_data " .
-        "WHERE pid = ? ", array($pid) );
-    $providerid = $find_provider['providerID'];
-  }
-  return $providerid;
-}
-
-// This writes a billing line item to the output page.
-//
-function echoLine($lino, $codetype, $code, $modifier, $ndc_info='',
-  $auth = TRUE, $del = FALSE, $units = NULL, $fee = NULL, $id = NULL,
-  $billed = FALSE, $code_text = NULL, $justify = NULL, $provider_id = 0, $notecodes='', $exclude ="0")
-{
-  global $code_types, $ndc_applies, $ndc_uom_choices, $justinit, $pid;
-  global $contraception, $usbillstyle, $hasCharges;
-
-  // If using line item billing and user wishes to default to a selected provider, then do so.
-  if($GLOBALS['default_fee_sheet_line_item_provider'] == 1 && $GLOBALS['support_fee_sheet_line_item_provider'] ==1 ) {
-    if ($provider_id == 0) {
-      $provider_id = 0 + findProvider();
-    }
-  }
-
-  if ($codetype == 'COPAY') {
-    if (!$code_text) $code_text = 'Cash';
-    if ($fee > 0) $fee = 0 - $fee;
-  }
-  if (! $code_text) {
-    $sqlArray = array();
-    $query = "select id, units, exclude_from_insurance_billing, code_text from codes where code_type = ? " .
-      " and " .
-      "code = ? and ";
-    array_push($sqlArray,$code_types[$codetype]['id'],$code);
-    if ($modifier) {
-      $query .= "modifier = ?";
-      array_push($sqlArray,$modifier);
-    } else {
-      $query .= "(modifier is null or modifier = '')";
-    }
-    $result = sqlQuery($query, $sqlArray);
-    $code_text = $result['code_text'];
-    $exclude = $result['exclude_from_insurance_billing'];
-
-    if (empty($units)) $units = max(1, intval($result['units']));
-    if (!isset($fee)) {
-      // Fees come from the prices table now.
-      $query = "SELECT prices.pr_price " .
-        "FROM patient_data, prices WHERE " .
-        "patient_data.pid = ? AND " .
-        "prices.pr_id = ? AND " .
-        "prices.pr_selector = '' AND " .
-        "prices.pr_level = patient_data.pricelevel " .
-        "LIMIT 1";
-      echo "\n<!-- $query -->\n"; // debugging
-      $prrow = sqlQuery($query, array($pid,$result['id']) );
-      $fee = empty($prrow) ? 0 : $prrow['pr_price'];
-    }
-  }
-  $fee = sprintf('%01.2f', $fee);
-  if (empty($units)) $units = 1;
-  $units = max(1, intval($units));
-  // We put unit price on the screen, not the total line item fee.
-  $price = $fee / $units;
-  $strike1 = ($id && $del) ? "<strike>" : "";
-  $strike2 = ($id && $del) ? "</strike>" : "";
-  echo " <tr>\n";
-  echo "  <td class='billcell'>$strike1" .
-    ($codetype == 'COPAY' ? xl($codetype) : $codetype) . $strike2;
-  //if the line to ouput is copay, show the date here passed as $ndc_info,
-  //since this variable is not applicable in the case of copay.
-  if($codetype == 'COPAY'){
-    echo "(".htmlspecialchars($ndc_info).")";
-    $ndc_info = '';
-  }
-  if ($id) {
-    echo "<input type='hidden' name='bill[".attr($lino)."][id]' value='$id'>";
-  }
-  echo "<input type='hidden' name='bill[".attr($lino)."][code_type]' value='".attr($codetype)."'>";
-  echo "<input type='hidden' name='bill[".attr($lino)."][code]' value='".attr($code)."'>";
-  echo "<input type='hidden' name='bill[".attr($lino)."][billed]' value='".attr($billed)."'>";
-  echo "</td>\n";
-  if ($codetype != 'COPAY') {
-    echo "  <td class='billcell'>$strike1" . text($code) . "$strike2</td>\n";
-  } else {
-    echo "  <td class='billcell'>&nbsp;</td>\n";
-  }
-  if ($billed) {
-    if (modifiers_are_used(true)) {
-      echo "  <td class='billcell'>$strike1" . text($modifier) . "$strike2" .
-        "<input type='hidden' name='bill[".attr($lino)."][mod]' value='".attr($modifier)."'></td>\n";
-    }
-    if (fees_are_used()) {
-      echo "  <td class='billcell' align='right'>" . text(oeFormatMoney($price)) . "</td>\n";
-      if ($codetype != 'COPAY') {
-        echo "  <td class='billcell' align='center'>" . text($units) . "</td>\n";
-      } else {
-        echo "  <td class='billcell'>&nbsp;</td>\n";
-      }
-    }
-    if (justifiers_are_used()) {
-      echo "  <td class='billcell' align='center'$usbillstyle>" . text($justify) . "</td>\n";
-    }
-
-    // Show provider for this line (if using line item billing).
-    if($GLOBALS['support_fee_sheet_line_item_provider'] ==1) {
-      echo "  <td class='billcell' align='center'>";
-    }
-    else
-    {
-      echo "  <td class='billcell' align='center' style='display: none'>";
-    }
-      genProviderSelect('', '-- '.xl("Default").' --', $provider_id, true);
-      echo "</td>\n";
-
-    if ($code_types[$codetype]['claim'] && !$code_types[$codetype]['diag']) {
-      echo "  <td class='billcell' align='center'$usbillstyle>" .
-        htmlspecialchars($notecodes, ENT_NOQUOTES) . "</td>\n";
-    }
-    else {
-      echo "  <td class='billcell' align='center'$usbillstyle></td>\n";
-    }
-    echo "  <td class='billcell' align='center'$usbillstyle><input type='checkbox'" .
-      ($auth ? " checked" : "") . " disabled /></td>\n";
-    echo "  <td class='billcell' align='center'><input type='checkbox'" .
-      " disabled /></td>\n";
-    if($GLOBALS['bill_to_patient'] ==1) {
-    echo "  <td class='billcell' align='center'$usbillstyle><input type='checkbox'" .
-      ($exclude ? " checked" : "") . " disabled /></td>\n";
-  }
-  }
-  else { // not billed
-    if (modifiers_are_used(true)) {
-      if ($codetype != 'COPAY' && ($code_types[$codetype]['mod'] || $modifier)) {
-        echo "  <td class='billcell'><input type='text' name='bill[".attr($lino)."][mod]' " .
-             "value='" . attr($modifier) . "' " .
-             "title='" . xla("Multiple modifiers can be separated by colons or spaces, maximum of 4 (M1:M2:M3:M4)") . "' " .
-             "value='" . attr($modifier) . "' size='" . attr($code_types[$codetype]['mod']) . "'></td>\n";
-      } else {
-        echo "  <td class='billcell'>&nbsp;</td>\n";
-      }
-    }
-    if (fees_are_used()) {
-      if ($codetype == 'COPAY' || $code_types[$codetype]['fee'] || $fee != 0) {
-        echo "  <td class='billcell' align='right'>" .
-          "<input type='text' name='bill[".attr($lino)."][price]' " .
-          "value='" . attr($price) . "' size='6'";
-        if (acl_check('acct','disc'))
-          echo " style='text-align:right'";
-        else
-          echo " style='text-align:right;background-color:transparent' readonly";
-        echo "></td>\n";
-        echo "  <td class='billcell' align='center'>";
-        if ($codetype != 'COPAY') {
-          echo "<input type='text' name='bill[".attr($lino)."][units]' " .
-          "value='" . attr($units) . "' size='2' style='text-align:right'>";
-        } else {
-          echo "<input type='hidden' name='bill[".attr($lino)."][units]' value='" . attr($units) . "'>";
-        }
-        echo "</td>\n";
-      } else {
-        echo "  <td class='billcell'>&nbsp;</td>\n";
-        echo "  <td class='billcell'>&nbsp;</td>\n";
-      }
-    }
-    if (justifiers_are_used()) {
-      if ($code_types[$codetype]['just'] || $justify) {
-        echo "  <td class='billcell' align='center'$usbillstyle ";
-        echo "title='" . xla("Select one or more diagnosis codes to justify the service") . "' >";
-        echo "<select name='bill[".attr($lino)."][justify]' onchange='setJustify(this)'>";
-        echo "<option value='" . attr($justify) . "'>" . text($justify) . "</option></select>";
-        echo "</td>\n";
-        $justinit .= "setJustify(f['bill[".attr($lino)."][justify]']);\n";
-      } else {
-        echo "  <td class='billcell'$usbillstyle>&nbsp;</td>\n";
-      }
-    }
-
-    // Show provider for this line (if using line item billing)
-    if($GLOBALS['support_fee_sheet_line_item_provider'] ==1) {
-      echo "  <td class='billcell' align='center'>";
-    }
-    else
-    {
-      echo "  <td class='billcell' align='center' style='display: none'>";
-    }
-      genProviderSelect("bill[$lino][provid]", '-- '.xl("Default").' --', $provider_id);
-      echo "</td>\n";
-
-    if ($code_types[$codetype]['claim'] && !$code_types[$codetype]['diag']) {
-      echo "  <td class='billcell' align='center'$usbillstyle><input type='text' name='bill[".attr($lino)."][notecodes]' " .
-        "value='" . htmlspecialchars($notecodes, ENT_QUOTES) . "' maxlength='50' size='8' /></td>\n";
-    }
-    else {
-      echo "  <td class='billcell' align='center'$usbillstyle></td>\n";
-    }
-    echo "  <td class='billcell' align='center'$usbillstyle><input type='checkbox' name='bill[".attr($lino)."][auth]' " .
-      "value='1'" . ($auth ? " checked" : "") . " /></td>\n";
-    echo "  <td class='billcell' align='center'><input type='checkbox' name='bill[".attr($lino)."][del]' " .
-      "value='1'" . ($del ? " checked" : "") . " /></td>\n";
-        //ADD THE NEW CHECKBOX "Bill to Patient EXCLUDE"
-    if($GLOBALS['bill_to_patient'] ==1) {
-      echo "  <td class='billcell' align='center'><input type='checkbox' name='bill[".attr($lino)."][exclude_from_insurance_billing]' " .
-        "value='1'" . ($exclude ? " checked" : "") . " /></td>\n";
-    }
-  }
-
-  echo "  <td class='billcell'>$strike1" . text($code_text) . "$strike2</td>\n";
-  echo " </tr>\n";
-
-  // If NDC info exists or may be required, add a line for it.
-  if ($codetype == 'HCPCS' && $ndc_applies && !$billed) {
-    $ndcnum = ''; $ndcuom = ''; $ndcqty = '';
-    if (preg_match('/^N4(\S+)\s+(\S\S)(.*)/', $ndc_info, $tmp)) {
-      $ndcnum = $tmp[1]; $ndcuom = $tmp[2]; $ndcqty = $tmp[3];
-    }
-    echo " <tr>\n";
-    echo "  <td class='billcell' colspan='2'>&nbsp;</td>\n";
-    echo "  <td class='billcell' colspan='6'>&nbsp;NDC:&nbsp;";
-    echo "<input type='text' name='bill[".attr($lino)."][ndcnum]' value='" . attr($ndcnum) . "' " .
-      "size='11' style='background-color:transparent'>";
-    echo " &nbsp;Qty:&nbsp;";
-    echo "<input type='text' name='bill[".attr($lino)."][ndcqty]' value='" . attr($ndcqty) . "' " .
-      "size='3' style='background-color:transparent;text-align:right'>";
-    echo " ";
-    echo "<select name='bill[".attr($lino)."][ndcuom]' style='background-color:transparent'>";
-    foreach ($ndc_uom_choices as $key => $value) {
-      echo "<option value='" . attr($key) . "'";
-      if ($key == $ndcuom) echo " selected";
-      echo ">" . text($value) . "</option>";
-    }
-    echo "</select>";
-    echo "</td>\n";
-    echo " </tr>\n";
-  }
-  else if ($ndc_info) {
-    echo " <tr>\n";
-    echo "  <td class='billcell' colspan='2'>&nbsp;</td>\n";
-    echo "  <td class='billcell' colspan='6'>&nbsp;" . xlt("NDC Data") . ": " . text($ndc_info) . "</td>\n";
-    echo " </tr>\n";
-  }
-
-  // For IPPF.  Track contraceptive services.
-  if (!$del) $contraception |= contraceptionClass($codetype, $code);
-
-  if ($fee != 0) $hasCharges = true;
-}
-
-// This writes a product (drug_sales) line item to the output page.
-//
-function echoProdLine($lino, $drug_id, $del = FALSE, $units = NULL,
-  $fee = NULL, $sale_id = 0, $billed = FALSE)
-{
-  global $code_types, $ndc_applies, $pid, $usbillstyle, $hasCharges;
-
-  $drow = sqlQuery("SELECT name FROM drugs WHERE drug_id = ?", array($drug_id) );
-  $code_text = $drow['name'];
-
-  $fee = sprintf('%01.2f', $fee);
-  if (empty($units)) $units = 1;
-  $units = max(1, intval($units));
-  // We put unit price on the screen, not the total line item fee.
-  $price = $fee / $units;
-  $strike1 = ($sale_id && $del) ? "<strike>" : "";
-  $strike2 = ($sale_id && $del) ? "</strike>" : "";
-  echo " <tr>\n";
-  echo "  <td class='billcell'>{$strike1}" . xlt("Product") . "$strike2";
-  echo "<input type='hidden' name='prod[".attr($lino)."][sale_id]' value='" . attr($sale_id) . "'>";
-  echo "<input type='hidden' name='prod[".attr($lino)."][drug_id]' value='" . attr($drug_id) . "'>";
-  echo "<input type='hidden' name='prod[".attr($lino)."][billed]' value='" . attr($billed) . "'>";
-  echo "</td>\n";
-  echo "  <td class='billcell'>$strike1" . text($drug_id) . "$strike2</td>\n";
-  if (modifiers_are_used(true)) {
-    echo "  <td class='billcell'>&nbsp;</td>\n";
-  }
-  if ($billed) {
-    if (fees_are_used()) {
-      echo "  <td class='billcell' align='right'>" . text(oeFormatMoney($price)) . "</td>\n";
-      echo "  <td class='billcell' align='center'>" . text($units) . "</td>\n";
-    }
-    if (justifiers_are_used()) {
-      echo "  <td class='billcell' align='center'$usbillstyle>&nbsp;</td>\n"; // justify
-    }
-    echo "  <td class='billcell' align='center'>&nbsp;</td>\n";             // provider
-    echo "  <td class='billcell' align='center'$usbillstyle>&nbsp;</td>\n"; // note codes
-    echo "  <td class='billcell' align='center'$usbillstyle>&nbsp;</td>\n"; // auth
-    echo "  <td class='billcell' align='center'><input type='checkbox'" .   // del
-      " disabled /></td>\n";
-  } else {
-    if (fees_are_used()) {
-      echo "  <td class='billcell' align='right'>" .
-        "<input type='text' name='prod[".attr($lino)."][price]' " .
-        "value='" . attr($price) . "' size='6'";
-      if (acl_check('acct','disc'))
-        echo " style='text-align:right'";
-      else
-        echo " style='text-align:right;background-color:transparent' readonly";
-      echo "></td>\n";
-      echo "  <td class='billcell' align='center'>";
-      echo "<input type='text' name='prod[".attr($lino)."][units]' " .
-        "value='" . attr($units) . "' size='2' style='text-align:right'>";
-      echo "</td>\n";
-    }
-    if (justifiers_are_used()) {
-      echo "  <td class='billcell'$usbillstyle>&nbsp;</td>\n"; // justify
-    }
-    echo "  <td class='billcell' align='center'>&nbsp;</td>\n"; // provider
-    echo "  <td class='billcell' align='center'$usbillstyle>&nbsp;</td>\n"; // note codes
-    echo "  <td class='billcell' align='center'$usbillstyle>&nbsp;</td>\n"; // auth
-    echo "  <td class='billcell' align='center'><input type='checkbox' name='prod[".attr($lino)."][del]' " .
-      "value='1'" . ($del ? " checked" : "") . " /></td>\n";
-    if($GLOBALS['bill_to_patient'] ==1) {
-      echo "  <td class='billcell' align='center'><input type='checkbox' name='bill[".attr($lino)."][exclude_from_insurance_billing]' " .
-        "value='1'" . ($del ? " checked" : "") . " /></td>\n";
-    }
-  }
-
-  echo "  <td class='billcell'>$strike1" . text($code_text) . "$strike2</td>\n";
-  echo " </tr>\n";
-
-  if ($fee != 0) $hasCharges = true;
-}
-
-
-// Compute a current checksum of Fee Sheet data from the database.
-//
-function visitChecksum($pid, $encounter, $saved=false) {
-  $rowb = sqlQuery("SELECT BIT_XOR(CRC32(CONCAT_WS(',', " .
-    "id, code, modifier, units, fee, authorized, provider_id, ndc_info, justify, billed" .
-    "))) AS checksum FROM billing WHERE " .
-    "pid = ? AND encounter = ? AND activity = 1",
-    array($pid, $encounter));
-  $rowp = sqlQuery("SELECT BIT_XOR(CRC32(CONCAT_WS(',', " .
-    "sale_id, inventory_id, prescription_id, quantity, fee, sale_date, billed" .
-    "))) AS checksum FROM drug_sales WHERE " .
-    "pid = ? AND encounter = ?",
-    array($pid, $encounter));
-  $ret = intval($rowb['checksum']) ^ intval($rowp['checksum']);
-  if (CHECKSUM_LOGGING) {
-    $comment = "Checksum = '$ret'";
-    $comment .= ", AJAX = " . (empty($_POST['running_as_ajax']) ? "false" : "true");
-    $comment .= ", Save = " . (empty($_POST['bn_save']) ? "false" : "true");
-    $comment .= ", Saved = " . ($saved ? "true" : "false");
-    newEvent("checksum", $_SESSION['authUser'], $_SESSION['authProvider'], 1, $comment, $pid);
-  }
-  return $ret;
-}
-
-// This is just for IPPF, to indicate if the visit includes contraceptive services.
-$contraception = 0;
 
 // Possible units of measure for NDC drug quantities.
 //
@@ -626,6 +210,7 @@ if (!$alertmsg && ($_POST['bn_save'] || $_POST['bn_save_close'])) {
           "id = ? AND billed = 0 AND activity = 1", array($code,$units,$fee,$modifier,$auth,$provid,$ndc_info,$justify,$notecodes, $exclude, $id) );
       }
     }
+////  End item already in database    
 
     // Otherwise it's a new item...
     else if (! $del) {
@@ -703,7 +288,7 @@ if (!$alertmsg && ($_POST['bn_save'] || $_POST['bn_save_close'])) {
     "pid = ? AND encounter = ?", array($main_provid,$main_supid,$main_order,$main_referr,$main_contract,$pid,$encounter) );
 
   // Save-and-Close is currently IPPF-specific but might be more generally
-  // useful.  It provides the ability to mark an encounter as billed
+  // Save-and-Close provides the ability to mark an encounter as billed
   // directly from the Fee Sheet, if there are no charges.
   if ($_POST['bn_save_close']) {
     $tmp1 = sqlQuery("SELECT SUM(ABS(fee)) AS sum FROM drug_sales WHERE " .
@@ -725,12 +310,6 @@ if (!$alertmsg && ($_POST['bn_save'] || $_POST['bn_save_close'])) {
     }
   }
 
-  // More IPPF stuff.
-  if (!empty($_POST['contrastart'])) {
-    $contrastart = $_POST['contrastart'];
-    sqlStatement("UPDATE patient_data SET contrastart = ?" .
-      " WHERE pid = ?", array($contrastart,$pid) );
-  }
 
   // Note: Taxes are computed at checkout time (in pos_checkout.php which
   // also posts to SL).  Currently taxes with insurance claims make no sense,
@@ -959,7 +538,7 @@ while ($prow = sqlFetchArray($pres)) {
     $ctkey = alphaCodeType($row['code_type']);
     if ($code_types[$ctkey]['nofs']) continue;
     echo "    <option value='" . attr($ctkey) . "|" .
-      attr($row['code']) . ':'. attr($row['modifier']) . "|'>" . text($row['code_text']) . "</option>\n";
+      attr($row['code']) . ':'. attr($row['modifier']) . "|'>" .text($row['code']) . ':'. text($row['modifier']) . '~'. text($row['code_text']) . "</option>\n";
   }
   echo "   </select>\n";
   echo "  </td>\n";
@@ -1359,33 +938,6 @@ echo "</b></span>\n";
 &nbsp;
 
 <?php
-// If applicable, ask for the contraceptive services start date.
-$trow = sqlQuery("SELECT count(*) AS count FROM layout_options WHERE " .
-  "form_id = 'DEM' AND field_id = 'contrastart' AND uor > 0");
-if ($trow['count'] && $contraception && !$isBilled) {
-  $date1 = substr($visit_row['date'], 0, 10);
-  // If admission or surgical, then force contrastart.
-  if ($contraception > 1 ||
-    strpos(strtolower($visit_row['pc_catname']), 'admission') !== false)
-  {
-    echo "   <input type='hidden' name='contrastart' value='" . attr($date1) . "' />\n";
-  }
-  else {
-    // echo "<!-- contraception = $contraception -->\n"; // debugging
-    $trow = sqlQuery("SELECT contrastart " .
-      "FROM patient_data WHERE " .
-      "pid = ? LIMIT 1", array($pid) );
-    if (empty($trow['contrastart']) || substr($trow['contrastart'], 0, 4) == '0000') {
-      $date0 = date('Y-m-d', strtotime($date1) - (60 * 60 * 24));
-      echo "   <select name='contrastart'>\n";
-      echo "    <option value='" . attr($date1) . "'>" . xlt('This visit begins new contraceptive use') . "</option>\n";
-      echo "    <option value='" . attr($date0) . "'>" . xlt('Contraceptive services previously started') . "</option>\n";
-      echo "    <option value=''>" . xlt('None of the above') . "</option>\n";
-      echo "   </select>\n";
-      echo "&nbsp; &nbsp; &nbsp;\n";
-    }
-  }
-}
 
 // If there is a choice of warehouses, allow override of user default.
 if ($prod_lino > 0) { // if any products are in this form
